isolate dropdown logic
